Separate concerns between what is related to the WASM call and the pre-receive logic
Currently the two are mixed together which will lead to code duplication if
we want to call a WASM module elsewhere.

The call convention between the library and PHP should not be a concern
of the code wanting to use a WASM module.

We expect one of those three messages from the Wasmtime wrapper library:
* success: `{"data": <wasm_module_output>}`
* user code error (e.g. unreachable code reached): `{"user_error": <error_message>}`
* internal error: `{"internal_error": <error_message>}`

Only `<wasm_module_output>` is under the control of the user writting
the WASM module, the rest is an internal convention. For example in the
case of a pre-receive hook WASM module we expect the response to look
something like this: `{"data":"{\\"rejection_message\\":null}"}`.

Part of story #31077: deploy a WebAssembly module to validate incoming references as a pre-receive hook

Change-Id: I1da7396c6f936c8d8e88c628b0b679ca2f36bac4