Disable Wamstime features not needed in Tuleap context
Wasmtime has a set of default features [0] enabling part
of the code that is not necessary for our usage:
* `async`: not used at the moment
* `cache` and `wat`: we expect to pre-process the WASM module the users
  provide us
* `jitdump` and `vtune`: we are unlikely to do this type of profiling in
  production

Keeping only the features we need result in a smaller library (and
compilation time) and reduce a bit our attack surface.

To test rebuild and redeploy the package on your dev (or manually build
it with `cargo-build-glibc217 --release` and symlink to it inside the
container).

Part of story #31077: deploy a WebAssembly module to validate incoming references
as a pre-receive hook

[0] https://github.com/bytecodealliance/wasmtime/blob/afb417920d0f4741edbbe25e4c10b4e6dcfb6ae0/crates/wasmtime/Cargo.toml#L58

Change-Id: I68ba8fc9bbd9a2d4c1d20a72df64fd3c94116195