Closes request #31531 turbo: 1.5.1 -> 1.8.8
Main release notes:
https://turbo.build/blog/turbo-1-8-0
https://turbo.build/blog/turbo-1-7-0
https://turbo.build/blog/turbo-1-6-0

Note that `turbo` complains explicitely about our usage of
`shared-workspace-lockfile=false` with a long warning.
We cannot switch to a shared workspace lockfile (for now) but we do not
rely on the features that would need it to work properly (e.g. `turbo
prune`) so it is not really an issue in our situation.

Change-Id: I8a74f8a1d0201a67c953ce585c92516cd067b87a