Closes request #31542 Stop building a "fat" Docker image to prepare the sources
Instead of building a "fat" Docker image we are now mounting `/nix`
inside the container and using `nix-shell`.
This saves a bit of time and it is closer to what we have in our
development environment. Also, it allows us to drop some of tricks that
were used to make it work (no more nix-env priority, setuid
binaries...).

Later this could also allow to share the Nix store between multiple CI
agents.

Change-Id: I316df0d4ed73cbfc3d6dcc632c2c4ce26c0f2345