feat: Store challenge of registration ceremony
Part of story #31173 register a new passkey

This commit introduce WebAuthnChallengeDao. It will allow to store challenge
generated for registration ceremony. The challenge need to be stored as the
second step of ceremony didn't know the challenge used for the first step.

*Testing:*

Fetch POST /webauthn/registration when connected. You will see a new
tuple in table webauthn_challenge with the id of your user.
The string stored in column challenge will be the same as the one in
challenge field of json returned by route

Change-Id: Ie33ee6db1b244e56fa8f37153d3e262da4436a24