Move MW 1.35 code into a subfolder
This is done in order to prepare the arrival of MW 1.39.
We will need to keep the MW 1.35 in order to be able to migrate
our existing MW 1.23 instances.

The MW 1.39 code will go in `/usr/share/mediawiki-tuleap-flavor/current/`
when we are ready to start using it.

To test you need to rebuild and deploy in your dev environment the package.
Also, do not forget to redeploy the nginx configuration.

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: If6d3500affa849aedc726584992ffa53c2fe3c82