feat: Have a multi-values autocompleter
story #31191 [pullrequest-overview] Add reviewers and make them updatable

This contribution implements a multi-values selection manager in
link-selector. The goal here is to have an complete autocompleter in
order to replace select2 where we need to select several values fetched
from REST on the fly.

In the next steps, I'll rename the lib "link-selector" in something more
generic, as this component is basically an autocompleter, and not
particularily a link-selector.

How to test:
- Go to the section "link-selector" in the section "Forms" of the
  tlp-doc.

--> There is a multi-users picker right below the link-selector
--> It has already a value selected
--> It will show some users if you start to type "jo" in its search
    input
--> Already selected values have the "selected" state in the dropdown
--> You can select several users
--> You can unselect all the users by clicking on the cross on the right
--> You can unselect a given user by clicking on its cross on its left

Change-Id: I44bb82debba7c6ad8c0e3b0ed7321f8c50d610bd