Migrating a MW 1.23 instance to MW Standalone now uses the CLI
This avoid the possible timeout issues that were encountered with the
REST calls.

The migration process is now: register instance -> migrate to 1.35 ->
deal with the existing permissions -> apply update for current MW
version (currently this step does nothing since our current version is
1.35).

The feature flag is now removed has it was added only to make sure the
users do not encounter the timeout issues when trying to migrate their
existing MW instances.

To test, do not forget to have a recent version of the MediaWiki Tuleap
flavor installed in your environment (or pull the latest version of the
image).

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: I851bd7335951aec1daa209c7b0256d25542bff86