Run MediaWiki CLI tools with PHP 8.1
MediaWiki 1.35 does not support officially PHP 8.1 but a lot of
compatibility patches have been backported to it. In any cases we will
need to have the CLI tools working with it since we are going to be forced
to keep MediaWiki 1.35 around to be able to upgrade from MediaWiki 1.23
(upgrading from MediaWiki 1.23 to something greater than MediaWiki 1.35
is not supported) and we cannot rely on an unsupported PHP version
forever.

Testing this change requires to run `tuleap-cfg site-deploy` to redeploy
the dynamic part of the LocalSettings files.

Part of story #26041: migrate mediwiki from 1.23 to 1.35.X

Change-Id: I2452b16a9762fb714fa6c717c2a513e56364bb67