Fixes request #31566: Document clipboard should not be shared accross projects
Issue was introduced in git #tuleap/stable/2c127691cd7f63768a04fb24b12a797cca2e1ee6.
`pinia-shared-state` was replaced with `@vueuse/core` because
`pinia-shared-state` does not allow the granularity we need and VueUse
is more likely to be maintained in the long run.

Determining which element need to be kept is now done at the element
level instead of the whole store. It makes the reactivity between the
different layers easier to comprehend when troubleshooting something.

Change-Id: I5bb1f3c68ea30dc8617290b7d8888cd928cfc9a3