refactor: Store pull_request_id as a number
part of story #31537 Add tags and make them updatable

The current pull-request id is retrieved from the route params as a
string and stored as is. We are forced to parse it as an int all the
time, and there are no reasons to keep it a string.

This contribution makes it a number and removes all the unecessary casts
here and there.

-- No functional change expected

Change-Id: Ica7de1b6d9fea3105c24dfbabb786015600f6713