Closes request #31574: Ignore vulnerability in a transitive dep of angular-gettext-cli/angular-gettext-loader
This contribution moves CVE-2021-3803 to the permanent ignore list.
It is located in one of the transitive deps of the Angular gettext
tooling which we only use at dev and build time. This ReDoS issue has
no real impact we have no easy way to fix it (proper fix is to remove
anything Angular related).

Change-Id: I176076ef3640a61650e7c98a8d17a14d0dca41b2