chore: e2e tuleap instance use regular startup process
The original `setup.sh` was made before we had a proper startup sequence
for Tuleap Docker Images. There is no good reasons to keep this
cumbersome shell script that spawn stuff around.

Not only we re-invent the wheel but we also completely bypass an
opportunity to dog food our docker images init sequence.

In order to be able to run stuff before supervisord, the startup of TCE
was aligned with the startup of TEE.

There should be no change in the execution of the test suite with this
patch.

Part of request #31584 It should be possible to run e2e test suite over regular Tuleap instance

Change-Id: I109dc80c5a127fcaa73f9663f0b83916731ae77a