feat: Display artifact progress in tooltip
When a semantic progress is set on a tracker then the tooltip of the
artifact always display the progress.

If there is an error in progress computation (ex: remaining effort >
initial effort) then an error is displayed to inform the user, like it
is done in roadmap tooltip.

If the progress is not configured, then no progress is displayed.

Part of story #26777: have artifact tooltips on roadmap

Change-Id: I51fccc97c7571c1c95b048599b04e276d2d7f92e