refactor: Single selection custom element
How to test:
- Lazybox in "single selection mode" (`is_multiple: false` in options)
now has a button that can be focused through the keyboard with "tab".
Pressing "enter" when the button is focused will clear the selection.
- Pressing "enter" is no longer enough to select items in the dropdown.
You must release "enter" (key up) to do the selection.
- No other functional change expected.

Notes: the button already existed but was a `<span>` and was not usable
by keyboard.

part of request #31583 Lazybox big clean-up

Change-Id: I5b5f73f8386f102ccf94ed01a77c52ac64ec9beb