Display warnings in the siteadministration when an outbound HTTP request is filtered
A warning is displayed on the siteadministration homepage and on the
system check events (which make possible to receive notifications for
it) when an outbound HTTP request is filtered. Also to make it easier
to find the information the filtered URLs are also logged in Tuleap
logfiles.

To test, even if right now nothing is filtered, you can try to contact
an endpoint returning you a 407 HTTp status code.

On instances where site administrators cannot control this, the
notifications can be disabled with
`tuleap config-set http_outbound_requests_filtered_alert never`.

Part of request #31580: Filter outbound HTTP requests

Change-Id: Ic984bd1f481c0dfc72a26d8a846a63d9b09b92c1