Fix request #31590 No more valid user value in modal is not cleared
Given a selectbox field bound to users (for example bound to project members),
and given an artifact having a user selected as the field value, if this user is
no more a project members, then this value is no more a valid one.

If we update this artifact with the Artifact UI, then this value is cleared and None is selected.
If we update this artifact with the Artifact modal, then this value is not cleared
even if None is selected and an error is raised.

Change-Id: I3a5b7f1709a4196910aa77b39f17403d7a6406cc