feat: inform admin that other semantics are in tooltip
When title, timeframe, or progress semantics are configured in the
current tracker, then in tooltip administration an alert informs the
user that they will be automatically displayed in the tooltip.

Note: extraction of the administration page in a mustache template or
conversion to BurningParrot will be done in a dedicated contribution.

Part of story #26777: have artifact tooltips on roadmap

Change-Id: I76804359913a35ca7ece0808ca7872a1626fb84f