Allow to exec bin when building packages
This is probablematic when building Rust code with crates having a
custom build step. The `tmpfs` volume type cannot be used for that
because does not apply the `exec` flag.

The `noexec` flag is kept when preparing the sources because it is the
place where we are the most likely to encounter weird stuff we want to
keep under control. For all the other steps it is code under our direct
control.

Part of request #31542: Stop building a "fat" Docker image to prepare the sources

Change-Id: Iaeb6ff57554e0191a95d9c00c77bd94aea694ca9