feat: Add route for WebAuthn account settings
Part of story #31173 register a new passkey

This commit introduces route for WebAuthn account settings with its
controller. In next commit there will be the mustache template.

*Testing:*

With WebAuthn plugin installed, go into account settings. There is a
WebAuthn tab, if you click on it, you land on an empty page with only
the header.

Change-Id: I3ee882fecc1e1ff4405b4afbd152f17365df411b