feat: kill switch for Smokescreen
We have made plenty of protection on new SSRF filtering proxy but we
have not thought about smokescreen misbehaving. If things goes full
wrong for some reasons, we should be able to give admins a way to bypass
smokescreen entirely.

How to test
-----------

1. Setup strict rules for your dev environment[1]
2. Ensure traffic is filtered
3. tuleap config-set filtering_proxy_usage disabled
4. Ensure traffic should no longer be filtered, no warning in system events
5. tuleap config-reset filtering_proxy_usage
6. Ensure traffice is filtered again

[1] https://docs.tuleap.org/administration-guide/system-administration/filtering-outbound-requests.html#adjusting-the-allow-and-deny-filter-lists

Part of: request #31580: Filter outbound HTTP requests

Change-Id: I73df28a7049212eadf38e06d626960f7b8cbf4c0