fix: long field name in tracker admin
As tracker administrator, create a field with the following label:
`Looooooooooooooooong`.
Then remove it.

A button should appear in the "Unused elements" palette with the label
that is not displayed outside of the button. An ellipsis is displayed
and either the description or the label of the button is displayed when
you hover it.

Closes request #31908: Long field name are displayed outside of button in administration

Change-Id: I7edc1627adff48228c39cd1e3d836dbf87df2f31