Hide SSRF filter bypass
This not supposed to be used, it should not be part of the public "API".
This flag is likely to be removed later once we have gained enough
confidence. In a "normal" situation disabling the filter is not
something that should be done, adding everything (`0.0.0.0/0,::/0`) to
the allow list is preferable since at least logs are collected.

Part of: request #31580: Filter outbound HTTP requests

Change-Id: If544077bd0cd6f7b7945e0d7a8458d8d0e6c8d7b