Fix request #31918 jira imported user fields must be bound to project members
Given a Tuleap instance with a lot of users (more than 5K active users), when doing
a project import from Jira, this will create multiple fields bound to all users.
With this amount of data, this will lead to huge loading in database when displaying
or editing an imported issue and could lead to a blank page.

We have to change it so fields are bound to project members instead.

Change-Id: I9c2e5e4a180c304fb5342af33cd173e0f697dda8