Fix request #31923 Broken link when more than 10 Jenkins jobs are triggered by one Jenkins hook
Given a Git repository that have a Jenkins webhook configured to trigger Git polling jobs,
and if there are more than 10 polling jobs triggered after a push, then there is a <x more>
returned by Jenkins after the 10th triggered job. This <x more> must not be clickable in the job list.

Change-Id: Ia8283f57c4352a85f65318a34980d028668ae95a