fix: Put back keyboard arrow navigation in lazybox dropdown
How to test:
In lazybox instances (tlp-doc, artifact link field in the modal,
reviewers in pullrequest overview), you can change the focus of the
dropdown items with the arrow up ↑ / down ↓ keys. If you reach the
beginning or the end, it cycles back to the other end.

Notes: In git #tuleap/stable/8b6e472d2210cda91506b088700e6bcbf2a80f8e,
keyboard arrow navigation was removed from lazybox (it was already a big
patch). This brings it back. The cycling behaviour is moved to a
dedicated library in the hope it will be reused: it could be applied to
swimlane navigation in taskboard, for example.

The "new item" button is also part of the navigation, but the search
input is excluded from it to avoid messing up the query. It is still
possible to go back to the search by navigating to the top element and
pressing shift+tab.

The editor examples now have more than one group of items to demonstrate
that the keyboard navigation "jumps" from one group to the next.

part of request #31583 Lazybox big clean-up

Change-Id: I20d576ae7a0fe912b397020504f200ac5fb2cc9c