fix: Storage of credenial public key in database
Part of story #31173 register a new passkey

The credential public key generated by passkeys along registration
ceremony is a binary string. Or this key is stored in text, resulting
on an empty string in database. To fix that, the column is now a blob.

*Testing:*

If you register your passkey in user preferences panel, then the column
'credential_public_key' in database is filled with some bytes.

Change-Id: Id32636044d0e17e7c5ca5035196f9aa931c80eb1