fix request #31941 Aggregate query is always run
When rendering a Table report with a field that can be aggregated
(selectbox), even when no aggregation function was defined, the query to
compute aggregate was always run.

On very large trackers (10k+), given that we pass the full list of
matching ids, that lead to extra large queries (when no other filters
reduce the amount of ids) that can take up to 50% of the rendering time
for nothing.

I took the opportunity to remove arguments to buildOrderedQuery that
were not used.

Change-Id: Ic8d1a5586bd2a8f26b7c9fc8f879f8c423a17f49