fix: Use event object instead of constant
When ldap is installed but git is not, the former is failing to listen
to `GerritCanMigrateChecker::GIT_EVENT_PLATFORM_CAN_USE_GERRIT` causing
the application to crash.

Introduced by git #tuleap/stable/d6aacaf3210b2548ac425166640270a339929525

To solve the issue we convert to an object instead of a constant,
because it is not possible to detect if a constant exist when it is
passed as an attribute parameter.

Took the opportunity to fix the namespace of GerritCanMigrateChecker and to remove Mockery from its test.

Part of request #31547: Simplify hook listening for plugins Part 2

Change-Id: I2cb1b9b51fec872d37a11e0b0df0721783d9ae78