refactor: Move register script into a subscript
Part of story #31173 register a new passkey

*No functional changes expected*

The account application has to be a module for future use with gettext. But cypress tests don't work with a module. So to fix that, the main package.json is for cypress and account application is a subscript with its own package.json with type module

Change-Id: Idd98cbfd846559315d4750896572f27d6cc20e44