Start using pnpm lockfile v6 format
This format is the one used by pnpm 8.x so we will need to move to it at
some point.

This new format is announced to be better for merge conflict resolution
so the merge constraint that was set on pnpm-lock.yml file was removed.

Part of request #31950: Bump to pnpm 8

Change-Id: I52351784268f99657ebac88511c36681f7dc9215