feat: XML Import Roadmap Filter
If you export a project with a filtered roadmap, and you import it as a
new project, then the imported roadmap is also filtered.

Note: preference name in the xml is changed from report_id to
filter_report_id so that it is more specific: less chances for headaches
the day we will introduce another config based on reports.

Part of story #26776: configure roadmap filters

Change-Id: I58fecfd27dbe42832188d904f151b0ef3ac888ef