fix: key creation date storage
Part of story #31174 see my registered passkeys

When we register a new passkey, its registration date is saved and then
each time we use it, its last use is saved too. To do that, there is
just default and on update rules in mysql table, problem : mysql
timezone is not the same as use timezone. To fix that, database store a
timestamp as an unsigned integer and not a datetime, and it's not the
database who set these values but the php script.

*Testing:*

Assuming mysql timezone is UTC and yours is UTC+2, before the patch if
your register a passkey at 11h, 9h will be displayed. Now it's 11h.

Change-Id: I5365818b4dbffbf4cccb1169434fe81a6ac52150