chore: Build Move artifact modal separately
No functional change expected in the "Move artifact" button. Edit an
artifact with the Artifact view, on the top left, there is an "Actions"
button with a dropdown and a "Move this artifact" button. When you click
it, the modal loads correctly and is translated. If it fails to load
(network offline), a codendi feedback is shown.

part of request #29264 Split the build of tracker plugin frontend assets
into multiple apps

Change-Id: Idb206cfbacaf21eafcb076a4cb51192a0343e121