refactor: Move LinkFieldController to Domain
No functional change expected in the Artifact link field behaviour in
the Artifact modal.

Notes: As it holds the most business logic, the Controller should be in
the Domain, not in the UI adapter. It should only handle Domain objects,
not presenters. Its job is to know everything the links field can do:
how to mark links for removal, how to retrieve existing links, how to
add new ones, how to retrieve possible parents, etc.

part of request #32274 Simplify Artifact link field

Change-Id: If2c0ca6823c280464290d9763156f50191f76dce