Close request #32278 Removed restricted users are not remove at project visibility switch
When switching from a project visibility that allows restricted users to Private without restricted,
then restricted users that are project administrators keep this access right, meaning that they can
continue to go to project and do some administration actions.

At project visibility switch to Private without restricted, we have to remove the project administration
right to restricted users. If this will remove the last administrator, then the project visibility switch
must be refused.

If the visibility change is triggered and a system events is logged,
then restricted adminsitrators are always removed if restricted users
are no more part of the project.

Note: Nothing is done for Public visibility without restricted as the
current code and implementation does nothing special for this
visibility. This can be handled in a dedicated commit if needed.

Change-Id: I42e2ef7c06e5e5e8af186a19139c532624b6e761