refactor: And & Or are Logical
The aim of this contribution is to better segregate Visitable
implementations. Currently we have the following inheritance:

Visitable
├ AndExpression
├ AndOperand
├ OrExpression
├ OrExpression
├ Comparison
│ ├ BetweenComparison
│ ├ GreaterThanOrEqualComparison
│ ├ NotEqualComparison
│ ├ LesserThanOrEqualComparison
│ ├ NotInComparison
│ ├ InComparison
│ ├ LesserThanComparison
│ ├ GreaterThanComparison
│ └ EqualComparison
└ Searchable
  ├ Metadata
  └ Field

This contribution introduces the `Logical` interface so that we now
have:

Visitable
├ Logical
│ ├ AndExpression
│ ├ AndOperand
│ ├ OrExpression
│ └ OrExpression
├ Comparison
│ ├ BetweenComparison
│ ├ GreaterThanOrEqualComparison
│ ├ NotEqualComparison
│ ├ LesserThanOrEqualComparison
│ ├ NotInComparison
│ ├ InComparison
│ ├ LesserThanComparison
│ ├ GreaterThanComparison
│ └ EqualComparison
└ Searchable
  ├ Metadata
  └ Field

Note: since we now have return types on getExpression/getOperand, psalm
is not happy about Term not having `accept` method (In
QueryBuilderVisitor for example). So now Term extends Visitable.

No functional changes

Preparation step for epic #32276: Filter on Artifact Links with TQL

Change-Id: I85f1fb96a21d2bc41c2b8b13efbd3d68c23b35f4