refactor: Enforce signature of SearchableVisitor
An interface is extracted and define the contract of the two methods
visitField and visitMetadata.

Since there is no generics in PHP, we rely on PHP annotations & psalm to
enforce the signature.

When we will introduce a new visitor, there will be less chance to miss
the implementation of one method (like it is currently the case for
`Visitor` implementations).

No functional change.

Preparation step for epic #32276: Filter on Artifact Links with TQL

Change-Id: I04160f172e9f6b9c9f9e350a3047895ebd82e4c9