refactor: Remove Visitor & Visitable interfaces
Those interfaces did not enforce return type so they are replaced with:
* SearchableVisitor[0] (field, metadata)
* ComparisonVisitor[1] (=, !=, <, >, <=, =>, IN, NOT IN, BETWEEN)
* LogicalVisitor 🆕 (and, or)

This forces us to better know what is coming from the parser: we no
longer have a random "Visitable". According to the grammar defined in
tql.peg.js, we will get an `OrExpression`, even for a simple query like
`field = 123`. This complexify a little bit some part of the parsing (we
cannot anymore have a common `visitExpression` since Or and And do not
behave the same way), but we gain typed visitors which is better IMHO.

No functional changes.

Part of story #32281: Search artifacts with/without parents via TQL

[0]: git #tuleap/stable/4bc1b7d17f9ca49941dcaeac35d01e94842a8b05
[1]: git #tuleap/stable/8f0b2cf3d44b5c89659f1e8c10a43ec0646fa1d6

Change-Id: Ia5cf6f3fda005aa0df110c60f4841696b630950e