chore: Argument type does not match the declared docblock
In Searchable and Logical, the acceptXxxx method has ambiguous
declaration: the signature of the method says it expects a
`VisitorParameters` while the docblock says it expects a `Parameters`.
Since we are relying on templates, we should use the one from the
docblock only.

Part of story #32281: Search artifacts with/without parents via TQL

Change-Id: I3ac3a6cfac9f7bccce5c4bd6b549518ff2bad7c0