feat: Improve error-handling when creating an artifact
How to test:
- When you create a new artifact with the modal (for example in Kanban)
and there is an error, the error message will appear at the top of the
modal in a fixed section (you no longer need to scroll to see it).
- When you click the submit button, it will be disabled while the REST
calls (file upload, artifact creation) are in-flight.
- When there is an error during file upload, it will also be shown at
the top of the modal in the fixed section.
- No functional change expected for artifact edition. Error message for
edition will be handled in a separate commit.

part of story #28823 [modal] create artifact
Also part of request #29640 Artifact modale should better highlight when
there are errors
Also part of request #10555 Artifact modal does not clearly says that
there are errors

Change-Id: I36595420714a4155ad6160961c5e7a018af46dbd